all magic bytes + extensions I could find
This shows, if we want to support all of these at some point, we need to
allow the same magic bytes for different formats. Sadly, we can't just
add this but must slowly migrate this to the registry. I'm against
letting `ImageMagick` just load all unrecognized files, so this should
be helpful.