make read and load import the IO library
This implies, that all IO libraries shouldn't export their load and save
functions, so they don't extend the FileIO functions anymore. This is a
bigger change, but it seems to solve our `Lib A, Lib B can load/save the
same format` problem nicely. It still works for libraries that overload,
but we should ultimately make the change.