Add LoaderError and WriterError exceptions
When throwing these errors in library code, FileIO will
emit an info message and then try the next registered loader / writer.

These are useful for marking specific failed assertions.  If the file
does not conform to what the loader expects, it can bail out and allow
the next reader to attempt to load the file.  At the same time,
legitimate programming errors will halt execution with an error message.