Improve precompilability with noninferrable filename
After #259, the result of `filename(q)` is not inferrable.
This uses "the kernel trick" to provide a single point where the result
must be inferred, and an `invokelatest` to prevent the abstract signature
from being inferred. It then precompiles the `::String` variant.
Consequently, with respect to latency this should get us back to where
we were before #259.