Major rewrite for correctness, performance
This package was started in 2015 (back in the Julia 0.3 or 0.4 days), and a lot
has changed since then. It's never really gotten a serious freshening.
This rewrite has several goals.

Improving the robustness of package/Module identification
---------------------------------------------------------

In modern versions of Julia, the package manager uses name/UUID combinations
to identify packages. This is far more robust and flexible than older strategies
for specifying packages. FileIO doesn't do this: it identifies modules by their
name only. We should adopt the new approach here: going forward (once the
deprecation period has passed and we release FileIO v2), all not-yet-loaded
modules must be specified by name/UUID.

There are some cases--often used in tests or transiently during development of
a new I/O package--where the handler *isn't* a registered package, and so there's
no UUID available. Currently we try to look up the module based on a `name::Symbol`.
It used to be that most modules were loaded into `Main`, then Julia switched
to `Base.__toplevel__`; currently we search both, since modules defined in the
REPL or tests might still live in `Main`.
Of course, even back in the old days, sub-modules could not be found in `Main`,
so the current system can't handle submodules.

To address the need for specifying modules that aren't packages, while
improving both correctness and flexibility, this PR allows you to
specify it by (duh) the module itself rather than the name of the module.

The combination of using either the module itself or a name/UUID combination
means that we can replace a lot of brittle & slow code. When we have the module,
we're done; when we have a name/UUID combination, we just call `Base.require`
to get the module. It even checks for us whether the module is already loaded.
End of story.

To help transition existing users to the new system, this has
"depwarn"-code to look for the module based on its name. It searches:

1. the currently-loaded modules
2. `Main`
3. The user's current `Pkg` environment

One key (breaking) difference is that this lookup is now done during `add_format`
rather than when the user tries to `load` or `save` a file. This is obviously
better for runtime efficiency, but it does change the point in the code where
an error occurs. One of the relatively changes to the tests addresses this change.

**Summary**: the new system is strictly more flexible than the old one, since we could
never previously support sub-modules. It is also strictly more correct since
the registry now specifies precisely what it means by `ImageIO`.
There is depwarn-code to help existing users transition, and the only known breakages
only concern the specific point in the code from which an error would be thrown.

Improving performance and reducing latency with better inferrability
--------------------------------------------------------------------

In the original design of this package, `load` and `save` were designed to be
specialized by packages. To allow format-specific dispatch, we encoded the
file format into the type system using types like `DataFormat{:PNG}`.
However, at a certain point we switched to calling module-specific
unexported `load` and `save` methods. As a consequence, we don't really
need to encode the format in the type system, we can just use a runtime
value. Indeed, the downside of using the type system is that having each
format be a separate type makes it impossible to infer types. This hurts the
runtime performance, increases latency due to unnecessary method specialization
by the compiler, and increases the risk of invalidation.

However, one way in which we may *under*-specialize is for the filename.
defined in `FilePathsBase`. That's a nice change, but this package does quite
a lot of manipulation based on file name, and having the type be non-inferrable
has some downsides.

Finally, several of the container types have historically been poorly-specified,
e.g., `const magic_list = Vector{Pair}()`.

This rewrite tries to straddle two goals: improving internal inferrability
while maintaining backwards compatibility. The strategy taken is to try to
wait until the last possible moment to construct non-inferrable objects---to wait
until the results are reported back to the caller.
In this rewrite, the data format is encoded internally just as a `Symbol`,
and the file is passed around as a separate object. This prevents one from
needing to specialize on the data format while preserving inferrability for the file.

There are a couple of minor changes to internal types, and this forced a couple of
changes to the tests. Most significantly, `File{fmt}` is no longer a concrete
type, because `File` got a second type-parameter to encode the filename type.
To prevent inference failures due to varying-length tuples, this also transitions
all magic bytes from `NTuple{N,UInt8}` to `Vector{UInt8}`.

As a case study, with the existing FileIO release, I get ~50us to load a
10x10 RGB png file. With this version, it's ~25us. It's remarkable that inference
can compete with I/O as a source of slowness, but there you have it.